Add go-to-doc feature
Currently the documentation is opened in the system's default
browser. It could get open in atom itself, 'cause atom is a chromium
but the needed plugin (browser-plus) is currently buggy on linux. It
doesn't only crash with an exception it causes Atom to freeze..
Therefore open links in external browser to avoid freezes!